server/requirements.txt
click==8.1.3
colorama==0.4.6
ffmpeg-python==0.2.0
Flask==2.2.3
Flask-Cors==3.0.10
future==0.18.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pytube==12.1.3
six==1.16.0
Werkzeug==2.2.3
